Fixed package name in akka http module
