+++ client-akka/src/test/scala/com/full360/prometheus/http/akka/AkkaHttpGaugeSpec.scala
@@ -19,44 +19,42 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.full360.prometheus.http.finatra
+package com.full360.prometheus.http.akka
 
 import com.full360.prometheus.BaseSpec
 
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 
-class AkkaHttpCounterSpec extends BaseSpec with ScalatestRouteTest with AkkaHttpCounter {
+class AkkaHttpGaugeSpec extends BaseSpec with ScalatestRouteTest with AkkaHttpGauge {
 
   val route =
-    counter {
+    gauge {
       pathSingleSlash {
         get {
+          registryShouldBe(
+            s"""# HELP ${gaugeNamespace}_$gaugeName $gaugeHelp
+               |# TYPE ${gaugeNamespace}_$gaugeName gauge
+               |${gaugeNamespace}_$gaugeName{method="get",path="/",} 1.0
+               |""".stripMargin
+          )
+
           complete("foo")
         }
       }
     }
 
-  "Counter metric" should provide {
-    "a counter DSL for Akka Http" which {
-      "increase the counter by 1" in {
+  "Gauge metric" should provide {
+    "a gauge DSL for Akka Http" which {
+      "increase and decrease by 1" in {
         Get() ~> route ~> check {
           responseAs[String] shouldBe "foo"
           registryShouldBe(
-            s"""# HELP ${counterNamespace}_$counterName $counterHelp
-               |# TYPE ${counterNamespace}_$counterName counter
-               |${counterNamespace}_$counterName{method="get",code="200",path="/",} 1.0
+            s"""# HELP ${gaugeNamespace}_$gaugeName $gaugeHelp
+               |# TYPE ${gaugeNamespace}_$gaugeName gauge
+               |${gaugeNamespace}_$gaugeName{method="get",path="/",} 0.0
                |""".stripMargin
           )
-
-          Get() ~> route ~> check {
-            registryShouldBe(
-              s"""# HELP ${counterNamespace}_$counterName $counterHelp
-                 |# TYPE ${counterNamespace}_$counterName counter
-                 |${counterNamespace}_$counterName{method="get",code="200",path="/",} 2.0
-                 |""".stripMargin
-            )
-          }
         }
       }
     }
